# api/openapi3.yml
openapi: "3.0.3"
info:
  version: 1.0.0
  title: Kids Activities Booking Platform - Backend API
  description: API documentation for the Kids Activities Booking Platform. Connecting parents with quality swimming, tutoring, and art classes for their children.
servers:
  - url: http://localhost:8000/v1
    description: Local environment
  - url: https://api-staging.kidsactivities.com/v1
    description: Staging environment
  - url: https://api.kidsactivities.com/v1
    description: Production environment
...
paths:
...
  /register/parent:
    post:
      summary: Register a new parent account
      operationId: RegisterParent
      tags:
        - auth
      requestBody:
        $ref: '#/components/requestBodies/RegisterParentRequest'
      responses:
        "201":
          $ref: '#/components/responses/RegisterResponse'
        "400":
          $ref: '#/components/responses/BadRequestResponse'
        default:
          $ref: '#/components/responses/InternalServerErrorResponse'
  /register/vendor:
    post:
      summary: Register a new vendor account
      operationId: RegisterVendor
      tags:
        - auth
      requestBody:
        $ref: '#/components/requestBodies/RegisterVendorRequest'
      responses:
        "201":
          $ref: '#/components/responses/RegisterVendorResponse'
        "400":
          $ref: '#/components/responses/BadRequestResponse'
        default:
          $ref: '#/components/responses/InternalServerErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    DefaultErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - message
          properties:
            message:
              type: string
            code:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
...
    UserRole:
      type: string
      enum:
        - parent
        - vendor
        - coach
        - admin
    UserStatus:
      type: string
      enum:
        - active
        - suspended
        - inactive
    VendorStatus:
      type: string
      enum:
        - pending
        - active
        - suspended
        - rejected
...
    BusinessType:
      type: string
      enum:
        - swimming_school
        - tutoring_center
        - art_studio
        - individual_coach
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        full_name:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        created_at:
          type: string
          format: date-time
...
    Vendor:
      type: object
      properties:
        id:
          type: integer
          format: int64
        business_name:
          type: string
        description:
          type: string
        business_type:
          $ref: '#/components/schemas/BusinessType'
        phone:
          type: string
        whatsapp:
          type: string
        address:
          type: string
        city:
          type: string
        district:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        logo:
          type: string
        cover_image:
          type: string
        photos:
          type: array
          items:
            type: string
        amenities:
          type: array
          items:
            type: string
        rating_avg:
          type: number
          format: double
        total_reviews:
          type: integer
        total_bookings:
          type: integer
        verified:
          type: boolean
        status:
          $ref: '#/components/schemas/VendorStatus'
...
  requestBodies:
...
    RegisterParentRequest:
      description: Request payload for parent registration
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - email
              - password
              - full_name
              - phone
            properties:
              email:
                type: string
              password:
                type: string
                minLength: 8
              full_name:
                type: string
              phone:
                type: string
              city:
                type: string
                default: Tangerang
              district:
                type: string
    RegisterVendorRequest:
      description: Request payload for vendor registration
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - email
              - password
              - full_name
              - phone
              - business_name
              - business_type
              - address
              - city
            properties:
              email:
                type: string
              password:
                type: string
                minLength: 8
              full_name:
                type: string
              phone:
                type: string
              business_name:
                type: string
              business_type:
                $ref: '#/components/schemas/BusinessType'
              address:
                type: string
              city:
                type: string
                default: Tangerang
              district:
                type: string
              whatsapp:
                type: string
...
  responses:
...
    BadRequestResponse:
      description: Bad request response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DefaultErrorResponse"
          example:
            error:
              message: "Validation failed"
              code: "VALIDATION_ERROR"
              details:
                - field: "email"
                  message: "Email is required"
                - field: "password"
                  message: "Password must be at least 8 characters"
...
    InternalServerErrorResponse:
      description: Internal server error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DefaultErrorResponse"
          example:
            error:
              message: "Internal server error"
              code: "INTERNAL_SERVER_ERROR"
...
    RegisterResponse:
      description: Successful registration response
      content:
        application/json:
          schema:
            type: object
            required:
              - data
            properties:
              data:
                type: object
                required:
                  - user
                properties:
                  user:
                    $ref: '#/components/schemas/User'
              message:
                type: string
    RegisterVendorResponse:
      description: Successful vendor registration response
      content:
        application/json:
          schema:
            type: object
            required:
              - data
            properties:
              data:
                type: object
                required:
                  - user
                  - vendor
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  vendor:
                    $ref: '#/components/schemas/Vendor'
              message:
                type: string
...